Move downsampler types to their own files
